Load real Point Cloud data from a PCD file representing a city block, and render the data.
This time we are using the pcl PointXYZI type. The "I" stands for intensity, which will now be an additional feature for each point in the cloud. In environment.cpp we create a new function called CityBlock that will be the same layout as the simpleHighway function. We use the point processor to load one of the car’s point clouds from a pcd file and then use the renderPointCloud function to view it.

The PNG images are the results of loading and running this code. If the color is not specified in the renderPointCloud function argument, it will default to using the intensity color coding. Looking around the pcd, we can see several cars parked along the sides of the road and a truck approaching to pass the ego car on the left side.