Implement Voxel Grid filtering with PCL
Voxel grid filtering will create a cubic grid and will filter the cloud by only leaving a single point per voxel cube, so the larger the cube length the lower the resolution of the point cloud.

A boxed region is defined and any points outside that box are removed.

The FilterCloud function returns a downsampled cloud with only points that were inside the region specified. The point resolution is much lower than the original, and it cropped everything inside the box points that were specified.

The PNG files show the results of the same fitlered point cloud from different angles.